+++ .gitignore
@@ -1,3 +1,3 @@
-.bundle
+.bundle/
 coverage/
 pkg/
