+++ demo/config.ru
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative "app"
+
+run App
