+++ demo/test/test_helper.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "simplecov"
+SimpleCov.start
+
+require "minitest/autorun"
+require "rack/test"
+require_relative "../app"
