+++ demo/views/index.erb
@@ -1,4 +1,3 @@
-
 <!doctype html>
 <html lang="en">
   <head>
@@ -10,7 +9,6 @@
     <link rel="stylesheet" href="/app.css">
     <script src="/app.js"></script>
   </head>
-
   <body>
     <div class="ui middle aligned center aligned grid">
       <div class="column">
