--- web-demo/config.ru
@@ -1,3 +1,0 @@
-require_relative 'app'
-
-run App
